# Performance optimizations - Dedicated functions for map and list diffs