--- a/mix.exs
+++ b/mix.exs
@@ -40,7 +40,8 @@ defmodule Jsonpatch.MixProject do
       {:credo, "~> 1.7.5", only: [:dev, :test], runtime: false},
       {:dialyxir, "~> 1.4", only: [:dev], runtime: false},
       {:ex_doc, "~> 0.31", only: [:dev], runtime: false},
-      {:jason, "~> 1.4", only: [:dev, :test]}
+      {:jason, "~> 1.4", only: [:dev, :test]},
+      {:benchee, "~> 1.4", only: [:dev]}
     ]
   end