Hi :wave: Thanks for an amazing library. While working on LiveVue patches, I've spotted some low-hanging fruits about performance of this package. So here I go 😄 

## Approach

First, I've created a small benchmark in `test/benchmark.exs` using Benchee (I've added it to mix.exs, only in dev). Then I've developed a new version of diff alongside existing one, and after being done I've replaced original code to create this PR. 

You can start benchmark by running `mix run test/benchmark.exs`, but currently it will only show numbers for existing implementation. To compare it to the old one we'd need to copy an old implementation into a separate module. 

## Overview of changes
My goal was to reduce the number of memory allocations. Each time I could skip creating a new object, I tried to do so. 

It turns out that a dedicated function for making a diff for maps and for lists is the best way, as it allows to write a highly-optimized code that almost entirely avoids memory allocations. 

## Results

In my tests new implementation is 2-3x faster, and consumes around 2.5 less memory. Log from benchee:

```
Benchmarking Faster JsonPatch with input Complex Lists - Task Management ...
Benchmarking Faster JsonPatch with input Complex Maps - E-commerce Order ...
Benchmarking Faster JsonPatch with input Deep Nesting - Configuration Tree ...
Benchmarking Faster JsonPatch with input Mixed Maps and Lists - Social Media Post ...
Benchmarking Original JsonPatch with input Complex Lists - Task Management ...
Benchmarking Original JsonPatch with input Complex Maps - E-commerce Order ...
Benchmarking Original JsonPatch with input Deep Nesting - Configuration Tree ...
Benchmarking Original JsonPatch with input Mixed Maps and Lists - Social Media Post ...
Calculating statistics...
Formatting results...

##### With input Complex Lists - Task Management #####
Name                         ips        average  deviation         median         99th %
Faster JsonPatch         93.72 K       10.67 μs    ±34.80%        9.67 μs       22.33 μs
Original JsonPatch       34.20 K       29.24 μs   ±315.50%       25.58 μs       69.88 μs

Comparison: 
Faster JsonPatch         93.72 K
Original JsonPatch       34.20 K - 2.74x slower +18.57 μs

Memory usage statistics:

Name                  Memory usage
Faster JsonPatch           8.27 KB
Original JsonPatch        21.97 KB - 2.66x memory usage +13.70 KB

**All measurements for memory usage were the same**

Reduction count statistics:

Name               Reduction count
Faster JsonPatch            1.10 K
Original JsonPatch          3.00 K - 2.73x reduction count +1.91 K

**All measurements for reduction count were the same**

##### With input Complex Maps - E-commerce Order #####
Name                         ips        average  deviation         median         99th %
Faster JsonPatch        100.00 K       10.00 μs   ±139.63%        8.75 μs       26.58 μs
Original JsonPatch       41.54 K       24.07 μs   ±283.08%       20.92 μs       49.00 μs

Comparison: 
Faster JsonPatch        100.00 K
Original JsonPatch       41.54 K - 2.41x slower +14.07 μs

Memory usage statistics:

Name                  Memory usage
Faster JsonPatch           7.34 KB
Original JsonPatch        16.30 KB - 2.22x memory usage +8.95 KB

**All measurements for memory usage were the same**

Reduction count statistics:

Name               Reduction count
Faster JsonPatch            0.98 K
Original JsonPatch          2.34 K - 2.39x reduction count +1.36 K

**All measurements for reduction count were the same**

##### With input Deep Nesting - Configuration Tree #####
Name                         ips        average  deviation         median         99th %
Faster JsonPatch         60.95 K       16.41 μs    ±26.29%       15.33 μs       28.21 μs
Original JsonPatch       24.94 K       40.10 μs    ±23.62%       38.13 μs       67.61 μs

Comparison: 
Faster JsonPatch         60.95 K
Original JsonPatch       24.94 K - 2.44x slower +23.69 μs

Memory usage statistics:

Name                  Memory usage
Faster JsonPatch          13.27 KB
Original JsonPatch        36.13 KB - 2.72x memory usage +22.86 KB

**All measurements for memory usage were the same**

Reduction count statistics:

Name               Reduction count
Faster JsonPatch            1.77 K
Original JsonPatch          4.70 K - 2.65x reduction count +2.93 K

**All measurements for reduction count were the same**
```

All tests are passing, so I have a pretty high confidence everything is working in the same way as before. Hope these changes will be useful!